backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "accelerate>=1.11.0",
    "difpy>=4.2.1",
    "fastapi[standard]>=0.116.2",
    "pillow>=11.3.0",
    "spandrel>=0.4.1",
    "tiler>=0.6.0",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "transformers>=4.57.1",
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
  { index = "pytorch-cpu", marker = "sys_platform == 'darwin'" }
]
torchvision = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
  { index = "pytorch-cpu", marker = "sys_platform == 'darwin'" }
]

[dependency-groups]
dev = [
    "ruff>=0.13.1",
]
